requirements.txt
# Core dependencies
crewai==0.186.1
embedchain>=0.1.128
langchain==0.3.27
langchain-community==0.3.27
langchain-ollama==0.2.0
langsmith>=0.3.18

# Configuration files
pyyaml>=6.0

# Note: json-repair conflicts resolved by removing graphrag and fnllm packages
# as they were not used in the project codebase

# Document processing
pillow==10.4.0
pytesseract==0.3.10
opencv-python==4.10.0.84
pdf2image==1.17.0

# Excel reporting and data analysis
pandas>=2.0.0
openpyxl>=3.1.0

# MCP support
mcp==1.0.0
httpx==0.27.0
aiohttp>=3.10.0

# Data handling
pydantic==2.10.3
python-dotenv==1.0.1
numpy==1.26.4
typing-extensions==4.12.2

# Utilities
rich==13.9.4
asyncio==3.4.3
tiktoken==0.9.0
# Note: protobuf version may show warnings due to tensorflow/grpcio-tools conflicts
# but the installation works correctly
protobuf>=6.31.1
